chore: clippy lints cleanup — redundant casts, derive Default, idiomatic iterators (v0.5.353)

+++ crates/chart/src/drawing/primitives_v2/annotations/price_label.rs
@@ -141,7 +141,7 @@ impl Primitive for PriceLabel {
             ctx.set_fill_color(CONTROL_POINT_FILL);
             ctx.set_stroke_width(1.5);
             ctx.begin_path();
-            ctx.arc(x, y, CONTROL_POINT_RADIUS as f64, 0.0, std::f64::consts::TAU);
+            ctx.arc(x, y, CONTROL_POINT_RADIUS, 0.0, std::f64::consts::TAU);
             ctx.fill();
             ctx.stroke();
         }
