+++ crates/chart/src/drawing/primitives_v2/annotations/triangle_up.rs
@@ -119,7 +119,7 @@ impl Primitive for TriangleUp {
             ctx.set_stroke_width(1.5);
             for (px, py) in [(x, y), (x2, y2)] {
                 ctx.begin_path();
-                ctx.arc(px, py, CONTROL_POINT_RADIUS as f64, 0.0, std::f64::consts::TAU);
+                ctx.arc(px, py, CONTROL_POINT_RADIUS, 0.0, std::f64::consts::TAU);
                 ctx.fill();
                 ctx.stroke();
             }
