+++ crates/chart/src/drawing/primitives_v2/fibonacci/retracement.rs
@@ -623,12 +623,12 @@ impl Primitive for FibRetracement {
             ctx.set_stroke_width(1.5);
 
             ctx.begin_path();
-            ctx.arc(x1, y1, CONTROL_POINT_RADIUS as f64, 0.0, std::f64::consts::TAU);
+            ctx.arc(x1, y1, CONTROL_POINT_RADIUS, 0.0, std::f64::consts::TAU);
             ctx.fill();
             ctx.stroke();
 
             ctx.begin_path();
-            ctx.arc(x2, y2, CONTROL_POINT_RADIUS as f64, 0.0, std::f64::consts::TAU);
+            ctx.arc(x2, y2, CONTROL_POINT_RADIUS, 0.0, std::f64::consts::TAU);
             ctx.fill();
             ctx.stroke();
         }
