+++ crates/chart/src/drawing/primitives_v2/pitchforks/inside_pitchfork.rs
@@ -647,7 +647,7 @@ impl Primitive for InsidePitchfork {
 
             for (px, py) in [(x1, y1), (x2, y2), (x3, y3)] {
                 ctx.begin_path();
-                ctx.arc(px, py, CONTROL_POINT_RADIUS as f64, 0.0, std::f64::consts::TAU);
+                ctx.arc(px, py, CONTROL_POINT_RADIUS, 0.0, std::f64::consts::TAU);
                 ctx.fill();
                 ctx.stroke();
             }
